llmb-run/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmb-run"
version = "1.2.0"
description = "Lightweight tool for automating large batches of LLM benchmarking workloads"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "rich>=10.0.0"
]

[project.scripts]
llmb-run = "llmb_run:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
